Bft ProposerSelector interface
